refactoring and implementing basic beam pipeline for counting
