--- default_config.yaml
@@ -1,11 +1,0 @@
-# Generate https://... links by default
-PREFERRED_URL_SCHEME: https
-
-# Don't sort results alphabetically
-JSON_SORT_KEYS: False
-
-# Set to false to prevent users from running arbitrary code
-DEBUG: False
-
-# UA to be used by
-# CUSTOM_UA: "wiki-topic (toolforge) -- <your-email-address>"
