+++ go.mod
@@ -1,0 +1,27 @@
+module github.com/htried/wiki-diff-privacy
+
+go 1.16
+
+replace github.com/htried/wiki-diff-privacy/wdp => ./wdp
+
+require (
+	cloud.google.com/go/bigquery v1.17.0 // indirect
+	cloud.google.com/go/datastore v1.5.0 // indirect
+	cloud.google.com/go/pubsub v1.10.2 // indirect
+	cloud.google.com/go/storage v1.14.0 // indirect
+	github.com/apache/beam v2.28.0+incompatible // indirect
+	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
+	github.com/golang/snappy v0.0.3 // indirect
+	github.com/google/differential-privacy/go v0.0.0-20210412184833-f3e565a14b7d // indirect
+	github.com/google/differential-privacy/privacy-on-beam v0.0.0-20210412184833-f3e565a14b7d // indirect
+	github.com/htried/wiki-diff-privacy/wdp v0.0.0-00010101000000-000000000000 // indirect
+	github.com/linkedin/goavro v2.1.0+incompatible // indirect
+	github.com/spf13/cobra v1.1.3 // indirect
+	golang.org/x/mod v0.4.2 // indirect
+	golang.org/x/net v0.0.0-20210414194228-064579744ee0 // indirect
+	golang.org/x/oauth2 v0.0.0-20210413134643-5e61552d6c78 // indirect
+	golang.org/x/sys v0.0.0-20210415045647-66c3f260301c // indirect
+	google.golang.org/api v0.44.0 // indirect
+	google.golang.org/genproto v0.0.0-20210415145412-64678f1ae2d5 // indirect
+	google.golang.org/grpc v1.37.0 // indirect
+)
