+++ server.go
@@ -3,65 +3,37 @@ package main
 import (
     "fmt"
     "net/http"
-    "io"
-    "io/ioutil"
-    "encoding/json"
+    "log"
     "html/template"
-  	"log"
-  	"math"
   	"strings"
-    "path/filepath"
-  	"strconv"
   	"time"
   	"os"
-  	"encoding/csv"
   	"reflect"
   	"context"
+  	"regexp"
 
-	"flag"
-	log "github.com/golang/glog"
-	"github.com/google/differential-privacy/privacy-on-beam/codelab"
 	"github.com/apache/beam/sdks/go/pkg/beam"
 
 	// The following import is required for accessing local files.
 	_ "github.com/apache/beam/sdks/go/pkg/beam/io/filesystem/local"
 
 	"github.com/apache/beam/sdks/go/pkg/beam/runners/direct"
+	"github.com/htried/wiki-diff-privacy/wdp"
+	
+	"github.com/google/differential-privacy/privacy-on-beam/pbeam"
+	"github.com/apache/beam/sdks/go/pkg/beam/transforms/stats"
+	"github.com/apache/beam/sdks/go/pkg/beam/io/textio"
 )
 
 func init() {
+	beam.RegisterType(reflect.TypeOf((*PageView)(nil)))
 	beam.RegisterType(reflect.TypeOf((*normalizeOutputCombineFn)(nil)))
 	beam.RegisterType(reflect.TypeOf(outputAccumulator{}))
+	beam.RegisterFunction(createPageViewsFn)
 	beam.RegisterFunction(convertToPairFn)
+	beam.RegisterFunction(extractPage)
 }
 
-type Article struct {
-	Name	string 	`json:"article"`
-	Views 	int 	`json:"views"`
-	Rank 	int 	`json:"rank"`
-}
-
-type Project struct {
-	Project 	string 		`json:"project"`
-	Access		string 		`json:"access"`
-	Year 		string 		`json:"year"`
-	Month		string 		`json:"month"`
-	Day 		string 		`json:"day"`
-	Articles 	[]Article 	`json:"articles"`
-}
-
-type ApiResponse struct {
-	Items 	[]Project 	`json:"items"`
-}
-
-type PageVars struct {
-	Lang 		string
-	MinCount 	int 
-	Epsilon		float64
-	Sensitivity int 
-	Alpha 		float64
-	PropWithin 	float64
-}
 
 type DPArticle struct {
 	Name	string
@@ -72,7 +44,12 @@ type DPArticle struct {
 	DoAgg 	string
 }
 
-var LanguageCodes = []string{"aa", "ab", "ace", "ady", "af", "ak", "als", "am", "an", "ang", "ar", "arc", "ary", "arz", "as", "ast", "atj", "av", "avk", "awa", "ay", "az", "azb", "ba", "ban", "bar", "bat-smg", "bcl", "be", "be-x-old", "bg", "bh", "bi", "bjn", "bm", "bn", "bo", "bpy", "br", "bs", "bug", "bxr", "ca", "cbk-zam", "cdo", "ce", "ceb", "ch", "cho", "chr", "chy", "ckb", "co", "cr", "crh", "cs", "csb", "cu", "cv", "cy", "da", "de", "din", "diq", "dsb", "dty", "dv", "dz", "ee", "el", "eml", "en", "eo", "es", "et", "eu", "ext", "fa", "ff", "fi", "fiu-vro", "fj", "fo", "fr", "frp", "frr", "fur", "fy", "ga", "gag", "gan", "gcr", "gd", "gl", "glk", "gn", "gom", "gor", "got", "gu", "gv", "ha", "hak", "haw", "he", "hi", "hif", "ho", "hr", "hsb", "ht", "hu", "hy", "hyw", "hz", "ia", "id", "ie", "ig", "ii", "ik", "ilo", "inh", "io", "is", "it", "iu", "ja", "jam", "jbo", "jv", "ka", "kaa", "kab", "kbd", "kbp", "kg", "ki", "kj", "kk", "kl", "km", "kn", "ko", "koi", "kr", "krc", "ks", "ksh", "ku", "kv", "kw", "ky", "la", "lad", "lb", "lbe", "lez", "lfn", "lg", "li", "lij", "lld", "lmo", "ln", "lo", "lrc", "lt", "ltg", "lv", "mai", "map-bms", "mdf", "mg", "mh", "mhr", "mi", "min", "mk", "ml", "mn", "mnw", "mr", "mrj", "ms", "mt", "mus", "mwl", "my", "myv", "mzn", "na", "nah", "nap", "nds", "nds-nl", "ne", "new", "ng", "nl", "nn", "no", "nov", "nqo", "nrm", "nso", "nv", "ny", "oc", "olo", "om", "or", "os", "pa", "pag", "pam", "pap", "pcd", "pdc", "pfl", "pi", "pih", "pl", "pms", "pnb", "pnt", "ps", "pt", "qu", "rm", "rmy", "rn", "ro", "roa-rup", "roa-tara", "ru", "rue", "rw", "sa", "sah", "sat", "sc", "scn", "sco", "sd", "se", "sg", "sh", "shn", "si", "simple", "sk", "sl", "sm", "smn", "sn", "so", "sq", "sr", "srn", "ss", "st", "stq", "su", "sv", "sw", "szl", "szy", "ta", "tcy", "te", "tet", "tg", "th", "ti", "tk", "tl", "tn", "to", "tpi", "tr", "ts", "tt", "tum", "tw", "ty", "tyv", "udm", "ug", "uk", "ur", "uz", "ve", "vec", "vep", "vi", "vls", "vo", "wa", "war", "wo", "wuu", "xal", "xh", "xmf", "yi", "yo", "za", "zea", "zh", "zh-classical", "zh-min-nan", "zh-yue", "zu"}
+type PageView struct {
+	ID 		string
+	Name 	string
+}
+
+var TopFiftyArticles [50]wdp.Article
 
 
 func Index(w http.ResponseWriter, r *http.Request) {
@@ -101,13 +78,20 @@ func PageViews(w http.ResponseWriter, r *http.Request) {
 		log.Print("error validating API arguments: ", err)
 	}
 
+	TopFiftyArticles, err = wdp.GetGroundTruth(vars.Lang)
+	if err != nil {
+		log.Print("error getting ground truth from API: ", err)
+	}
+
 	var yesterday = time.Now().AddDate(0, 0, -1).Format("2006_01_02")
 	err = wdp.RemoveOldContents(yesterday, "data/")
 	if err != nil {
 		log.Print("error removing contents of data folder: ", err)
 	}
-	f, err = os.Open(fmt.Sprintf("./data/synthetic_data_%s_%s.csv", vars.Lang, yesterday))
-	defer f.Close()
+
+	fname := fmt.Sprintf("./data/synthetic_data_%s_%s.csv", vars.Lang, yesterday)
+	outname := fmt.Sprintf("./data/output_%s_%s.csv", vars.Lang, yesterday)
+	_, err = os.Stat(fname)
 	if os.IsNotExist(err) {
 		err = wdp.InitializeSyntheticData(yesterday, vars.Lang)
 		if err != nil {
@@ -118,20 +102,31 @@ func PageViews(w http.ResponseWriter, r *http.Request) {
 	} else {
 		// open file and run beam pipeline here
 		beam.Init()
-
 		p := beam.NewPipeline()
 		s := p.Root()
 
-		pvs := wdp.ReadInput(s, f)
-		rawOutput := wdp.CountPageViews(s, pvs)
-		// dpOutput := wdp.PrivateCountPageViews(s, pvs)
+		pvs := readInput(s, fname)
+		log.Print("read")
+		pageviews := beam.ParDo(s, extractPage, pvs)
+		log.Print("extracted")
+		counted := stats.Count(s, pageviews)
+		log.Print("counted")
+		formatted := beam.ParDo(s, func(page string, count int) string {
+			return fmt.Sprintf("%s: %v", page, count)
+		}, counted)
+		log.Print("formatted")
+		textio.Write(s, outname, formatted)
+		log.Print("written")
 
-		wdp.WriteOutput(s, rawOutput, fmt.Sprintf("./data/output_%s_%s.csv", vars.Lang, yesterday)
-		// wdp.WriteOutput(s, dpOutput, fmt.Sprintf("./data/output_%s_%s.csv", vars.Lang, yesterday)
+		// rawOutput := countPageViews(s, pvs)
+		// dpOutput := privateCountPageViews(s, pvs)
+
+		// writeOutput(s, rawOutput, outname)
+		// wdp.WriteOutput(s, dpOutput, outname)
 		// Execute pipeline.
-		err := direct.Execute(context.Background(), p)
+		_, err = direct.Execute(context.Background(), p)
 		if err != nil {
-			log.Exitf("Execution of pipeline failed: %v", err)
+			log.Print("Execution of pipeline failed: %v", err)
 		}
 	}
 }
@@ -139,7 +134,6 @@ func PageViews(w http.ResponseWriter, r *http.Request) {
 
 // bind functions to paths and start listener
 func main() {
-
 	// undo at the end
     // http.HandleFunc("/", Index)
     // http.HandleFunc("/api/v1/pageviews", PageViews)
@@ -148,37 +142,82 @@ func main() {
 }
 
 
+
+
+func createPageViewsFn(line string, emit func(PageView)) error {
+	// Skip the column headers line
+	notHeader, err := regexp.MatchString("[0-9]", line)
+	if err != nil {
+		return err
+	}
+	if !notHeader {
+		return nil
+	}
+
+	cols := strings.Split(line, "|")
+	if len(cols) != 2 {
+		return fmt.Errorf("got %d number of columns in line %q, expected 2", len(cols), line)
+	}
+	id := cols[0]
+	name := cols[1]
+	emit(PageView{
+		ID:		id,
+		Name: 	name,
+	})
+	return nil
+}
+
+// readInput reads from a .csv file detailing page views in the form
+// of "id, name" and returns a PCollection of Visit structs.
+
+// from the privacy on beam codelab tutorial
+func readInput(s beam.Scope, input string) beam.PCollection {
+	s = s.Scope("readInput")
+	lines := textio.Read(s, input)
+	return beam.ParDo(s, createPageViewsFn, lines)
+}
+
+func writeOutput(s beam.Scope, output beam.PCollection, outputTextName string) {
+	s = s.Scope("writeOutput")
+	output = beam.ParDo(s, convertToPairFn, output)
+	formattedOutput := beam.Combine(s, &normalizeOutputCombineFn{}, output)
+	log.Print("combined: ", formattedOutput)
+	textio.Write(s, outputTextName, formattedOutput)
+}
+
+
 // functions and types for aggregation in beam
 // from codelab example, with light changes
 type pair struct {
 	K string
-	V float64
+	V int
 }
 
-func convertToPairFn(k string, v beam.V) (pair, error) {
-	switch v := v.(type) {
-	case int:
-		return pair{K: k, V: float64(v)}, nil
-	case int64:
-		return pair{K: k, V: float64(v)}, nil
-	case float64:
-		return pair{K: k, V: v}, nil
-	default:
-		return pair{}, fmt.Errorf("expected int, int64 or float64 for value type, got %v", v)
-	}
+func convertToPairFn(k string, v int) (pair, error) {
+
+// func convertToPairFn(k string, v beam.V) (pair, error) {
+	// switch v := v.(type) {
+	// case int:
+	// 	return pair{K: k, V: float64(v)}, nil
+	// case int64:
+	// 	return pair{K: k, V: float64(v)}, nil
+	// case float64:
+	return pair{K: k, V: v}, nil
+	// default:
+	// 	return pair{}, fmt.Errorf("expected int, int64 or float64 for value type, got %v", v)
+	// }
 }
 
 type outputAccumulator struct {
-	PageToValue map[string]float64
+	PageToValue map[string]int
 }
 
 type normalizeOutputCombineFn struct{}
 
 func (fn *normalizeOutputCombineFn) CreateAccumulator() outputAccumulator {
-	pageToValue := make(map[string]float64)
-	topFiftyArticles, _ := GetGroundTruth(lang)
-	for _, article := range topFiftyArticles {
-		pageToValue[article] = 0
+	pageToValue := make(map[string]int)
+	for _, article := range TopFiftyArticles {
+		pageToValue[article.Name] = 0
 	}
 	return outputAccumulator{pageToValue}
 }
@@ -197,10 +236,41 @@ func (fn *normalizeOutputCombineFn) MergeAccumulators(a, b outputAccumulator) ou
 	return a
 }
 
-func (fn *normalizeOutputCombineFn) ExtractOutput(a outputAccumulator) string {
+func (fn *normalizeOutputCombineFn) extractOutput(a outputAccumulator) string {
 	var lines []string
 	for k, v := range a.PageToValue {
-		lines = append(lines, fmt.Sprintf("%d %f", k, v))
+		lines = append(lines, fmt.Sprintf("%s, %v", k, v))
 	}
 	return strings.Join(lines, "\n")
+}
+
+
+func countPageViews(s beam.Scope, col beam.PCollection) beam.PCollection {
+	s = s.Scope("countPageViews")
+	pageviews := beam.ParDo(s, extractPage, col)
+	viewsPerPage := stats.Count(s, pageviews)
+	return viewsPerPage
+}
+
+func privateCountPageViews(s beam.Scope, col beam.PCollection, epsilon float64, sensitivity int) beam.PCollection {
+	s = s.Scope("countPageViews")
+
+	spec := pbeam.NewPrivacySpec(epsilon, float64(sensitivity))
+	pCol := pbeam.MakePrivateFromStruct(s, col, spec, "id")
+
+	pageviews := pbeam.ParDo(s, extractPage, pCol)
+	viewsPerPage := pbeam.Count(s, pageviews, pbeam.CountParams{
+		MaxPartitionsContributed:	1, // In the scheme I've constructed, each visitor visits once per day
+		MaxValue: 					1, // And they can visit a maximum of one page
+	})
+	return viewsPerPage
+}
+
+func extractPage(p PageView) string {
+	reg, err := regexp.Compile(`[^a-zA-Z0-9\_]+`)
+    if err != nil {
+        log.Print("error with creating regex: ", err)
+    }
+    
+	return reg.ReplaceAllString(p.Name, "")
 }
