+++ wdp/go.mod
@@ -1,0 +1,3 @@
+module github.com/htried/wiki-diff-privacy/wdp
+
+go 1.16
